testing-projects/Django-MLM/requirements.txt
crispy-bootstrap5==0.7
Django==4.2.2
django-crispy-forms==2.0
django-environ==0.10.0
django-import-export==3.2.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.1.0
multidict==6.0.4
phonenumbers==8.13.14
Pillow==9.5.0
PyJWT==2.7.0
sqlparse==0.4.4
twilio==8.3.0
gunicorn==21.2.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
shortuuid==1.0.11